renamed style sheets to work correctly
