+++ application/src/main/resources/static/js/react-client/src/appStyle.module.scss
