+++ application/src/main/resources/static/js/react-client/src/components/BooksApi/bookStyle.module.scss
@@ -1,0 +1,17 @@
+// .booksstyle{
+//     margin: 40px 40px;
+//     font-family: 'Lucida Sans', 'Lucida Sans Regular', 'Lucida Grande', 'Lucida Sans Unicode', Geneva, Verdana, sans-serif;
+//     font-weight: 600;
+//     color: var(--blue);
+//     display: flex;
+//     flex-direction: column;
+//     border: #f56200;
+//     border-style: solid;
+//     border-radius: 15px;
+//     width: 50%;
+//     margin-left: 15px;
+//     padding-left: 50px;
+//     padding-right: 10px;
+//     padding-bottom: 30px;
+//     background: greenyellow;
+// }
