+++ application/src/main/resources/static/js/react-client/src/components/kidsActivitiesAPI/activities.js
@@ -1,5 +1,5 @@
 import React from "react";
-import style from './style.module.scss'
+import style from './kidsStyle.module.scss'
 
 const ActivitiesAPI = () => {
