+++ application/src/main/resources/static/js/react-client/src/components/scores/scoresStyle.module.scss
