+++ application/src/main/resources/static/js/react-client/src/pages/contact-screen/index.js
@@ -1,4 +1,5 @@
 import React from "react";
+import style from './contactStyle.module.scss';
 
 const ContactScreen = () => {
     return (
